separate primitives
